Extract the Selection class
